fix: revert scope creep, keep only version sync changes
Revert all source code, test, and generated files to main's version.
Keep only the core version discrepancy fix:
- VERSION: 0.3.5 -> 0.3.6
- version-integrity.yml: add cli-npm and VERSION triggers
- verify-version-sync.sh: check cli-npm and VERSION file
- test-version-sync.sh: new test for verification script
