+++ Cargo.toml
@@ -52,7 +52,7 @@ include = [
 ]
 
 [dependencies]
-tempfile = { workspace = true, optional = true }
+tempfile = { workspace = true }
 # Serialization
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -110,7 +110,6 @@ console_error_panic_hook = { version = "0.1", optional = false }
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
-tempfile = { workspace = true }
 proptest = "1.6.0"
 # CLI testing
 assert_cmd = "2.0.16"
@@ -244,7 +243,7 @@ embed-openai = ["dep:reqwest"]
 embed-voyage = ["dep:reqwest"]
 
 # ANN backends (ADR-0068)
-ann-hnsw = ["dep:hnsw_rs", "dep:tempfile"]
+ann-hnsw = ["dep:hnsw_rs"]
 ann-lsh = []
