+++ src/lib.rs
@@ -64,8 +64,6 @@ mod framework_bridge;
 pub mod framework_builder;
 mod framework_events;
 pub mod framework_events_ce;
-#[cfg(not(target_arch = "wasm32"))]
-mod framework_export;
 mod framework_graph_rag;
 mod framework_metrics;
 mod framework_namespaces;
@@ -127,7 +125,6 @@ pub mod persistence {
     //! Enable the "persistence" feature for full libSQL-backed persistence.
 
     use crate::error::Result;
-    use crate::hyperdim::HVec10240;
     use crate::singularity::Concept;
 
     /// Stub persistence type when persistence feature is disabled.
@@ -138,54 +135,53 @@ pub mod persistence {
     pub use crate::singularity::ConceptVersion;
 
     impl Persistence {
-        pub async fn save_concept(&self, _concept: &Concept) -> Result<()> {
+        pub async fn save_concept(&self, _ns: &str, _concept: &Concept) -> Result<()> {
             Ok(())
         }
 
-        pub async fn save_concepts(&self, _concepts: &[Concept]) -> Result<()> {
+        pub async fn save_concepts(&self, _ns: &str, _concepts: &[Concept]) -> Result<()> {
             Ok(())
         }
 
-        pub async fn load_concept(&self, _id: &str) -> Result<Option<Concept>> {
+        pub async fn load_concept(&self, _ns: &str, _id: &str) -> Result<Option<Concept>> {
             Ok(None)
         }
 
-        pub async fn load_all_concepts(&self) -> Result<Vec<Concept>> {
+        pub async fn load_all_concepts(&self, _ns: &str) -> Result<Vec<Concept>> {
             Ok(Vec::new())
         }
 
-        pub async fn delete_concept(&self, _id: &str) -> Result<()> {
+        pub async fn delete_concept(&self, _ns: &str, _id: &str) -> Result<()> {
             Ok(())
         }
 
-        pub async fn save_association(&self, _from: &str, _to: &str, _strength: f32) -> Result<()> {
+        pub async fn save_association(
+            &self,
+            _ns: &str,
+            _from: &str,
+            _to: &str,
+            _strength: f32,
+        ) -> Result<()> {
             Ok(())
         }
 
         pub async fn save_associations(
             &self,
+            _ns: &str,
             _associations: &[(String, String, f32)],
         ) -> Result<()> {
             Ok(())
         }
 
-        pub async fn load_associations(&self, _id: &str) -> Result<Vec<(String, f32)>> {
+        pub async fn load_associations(&self, _ns: &str, _id: &str) -> Result<Vec<(String, f32)>> {
             Ok(Vec::new())
         }
 
-        pub async fn concept_count(&self, _ns: &str) -> Result<usize> {
-            Ok(0)
-        }
-
-        pub async fn association_count(&self, _ns: &str) -> Result<usize> {
-            Ok(0)
-        }
-
-        pub async fn delete_association(&self, _from: &str, _to: &str) -> Result<()> {
+        pub async fn delete_association(&self, _ns: &str, _from: &str, _to: &str) -> Result<()> {
             Ok(())
         }
 
-        pub async fn clear_concept_associations(&self, _id: &str) -> Result<()> {
+        pub async fn clear_concept_associations(&self, _ns: &str, _id: &str) -> Result<()> {
             Ok(())
         }
 
@@ -232,6 +228,7 @@ pub mod persistence {
 
         pub async fn get_concept_history(
             &self,
+            _ns: &str,
             _id: &str,
             _limit: usize,
         ) -> Result<Vec<ConceptVersion>> {
@@ -242,15 +239,23 @@ pub mod persistence {
             Ok(0)
         }
 
-        pub async fn save_index(&self, _id: &str, _data: &[u8]) -> Result<()> {
+        pub async fn save_index(&self, _ns: &str, _id: &str, _data: &[u8]) -> Result<()> {
             Ok(())
         }
 
-        pub async fn load_index(&self, _id: &str) -> Result<Option<Vec<u8>>> {
+        pub async fn load_index(&self, _ns: &str, _id: &str) -> Result<Option<Vec<u8>>> {
             Ok(None)
         }
 
         pub async fn apply_migrations(&self, _target_version: i64) -> Result<()> {
+            Ok(())
+        }
+
+        pub async fn list_namespaces(&self) -> Result<Vec<String>> {
+            Ok(Vec::new())
+        }
+
+        pub async fn clear_namespace(&self, _ns: &str) -> Result<()> {
             Ok(())
         }
     }
@@ -260,7 +265,7 @@ pub mod prelude {
     pub use crate::bridge_retrieval::BridgeRetrieval;
     pub use crate::bundle::BundleAccumulator;
     pub use crate::error::{MemoryError, Result};
-    pub use crate::framework::{ChaoticSemanticFramework, MAX_IMPORT_SIZE};
+    pub use crate::framework::ChaoticSemanticFramework;
     pub use crate::framework_builder::FrameworkBuilder;
     pub use crate::framework_events::MemoryEvent;
     pub use crate::hyperdim::HVec10240;
