+++ src/reservoir_tests.rs
@@ -62,10 +62,51 @@ fn step_mathematical_correctness() {
     let manual_norm: f64 = out
         .state
         .iter()
-        .map(|&x| (x as f64).powi(2))
+        .map(|&x| (x as f64) * (x as f64))
         .sum::<f64>()
         .sqrt();
     assert!((out.state_norm - manual_norm).abs() < 1e-6);
+}
+
+#[test]
+#[allow(clippy::float_cmp)]
+fn norm_calculation_consistency() {
+    // 1. Verify direct multiplication matches powi(2) for range of values
+    // Using to_bits() to bypass clippy::float_cmp while ensuring bit-identical results
+    for i in -100..100 {
+        let x = i as f64 * 0.01;
+        assert_eq!((x * x).to_bits(), x.powi(2).to_bits());
+    }
+
+    // 2. Verify reservoir state_norm consistency over many steps
+    // Use a small reservoir and update_stride for test speed
+    let mut r = Reservoir::new_seeded(10, 100, 42).unwrap();
+    // update_stride is 32 by default. Run 100 steps to trigger 3+ full re-calculations.
+    let input = vec![0.1; 10];
+
+    for i in 0..100 {
+        let out = r.step(&input).unwrap();
+
+        // Manual calculation of norm from the actual state
+        let manual_norm_sq: f64 = out.state.iter().map(|&x| f64::from(x) * f64::from(x)).sum();
+        let manual_norm = manual_norm_sq.sqrt();
+
+        // Precision drift should be minimal (within 1e-9 for f64 accumulation)
+        assert!(
+            (out.state_norm - manual_norm).abs() < 1e-9,
+            "Norm mismatch at step {}: internal={}, manual={}",
+            i,
+            out.state_norm,
+            manual_norm
+        );
+
+        if r.update_phase == 0 {
+            // This was a full re-calculation step (or it will be at the start of next step if we check before increment)
+            // Actually in code: `if update_phase == 0 { re-calculate }` happens BEFORE `update_phase` is incremented.
+            // But we already called step, so `r.update_phase` was just incremented.
+            // If it's now 1, the previous step was phase 0.
+        }
+    }
 }
 
 #[test]
