+++ src/wasm.rs
@@ -1,7 +1,7 @@
 //! WASM bindings for chaotic semantic memory
 
 use bincode::Options;
-use js_sys::{Array, Uint8Array};
+use js_sys::{Array, Float32Array, Uint8Array};
 use tracing::warn;
 use wasm_bindgen::JsCast;
 use wasm_bindgen::prelude::*;
@@ -9,8 +9,7 @@ use wasm_bindgen::prelude::*;
 pub(crate) use crate::export_payload::{BinaryExportPayload, ExportPayload, unix_now_secs};
 pub(crate) use crate::framework::{ChaoticSemanticFramework, MAX_IMPORT_SIZE};
 pub(crate) use crate::hyperdim::HVec10240;
-pub(crate) use crate::singularity::Concept;
-
+pub(crate) use crate::wasm_ext::{concept_to_js_value, to_js_error};
 
 #[wasm_bindgen(start)]
 pub fn initialize_wasm() {
@@ -245,6 +244,108 @@ impl WasmFramework {
         Ok(results)
     }
 
+    /// Get framework metrics snapshot
+    pub async fn metrics_snapshot(&self) -> Result<JsValue, JsValue> {
+        let metrics = self.framework.metrics_snapshot().await;
+        let obj = js_sys::Object::new();
+        js_sys::Reflect::set(
+            &obj,
+            &"concepts_injected_total".into(),
+            &(metrics.concepts_injected_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"associations_created_total".into(),
+            &(metrics.associations_created_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"probes_total".into(),
+            &(metrics.probes_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"avg_probe_latency_ms".into(),
+            &metrics.avg_probe_latency_ms.into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"cache_hits_total".into(),
+            &(metrics.cache_hits_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"cache_misses_total".into(),
+            &(metrics.cache_misses_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"cache_evictions_total".into(),
+            &(metrics.cache_evictions_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"reservoir_steps_total".into(),
+            &(metrics.reservoir_steps_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"avg_reservoir_step_latency_us".into(),
+            &metrics.avg_reservoir_step_latency_us.into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"reservoir_nodes_active".into(),
+            &(metrics.reservoir_nodes_active as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"persist_ops_total".into(),
+            &(metrics.persist_ops_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"avg_persist_latency_ms".into(),
+            &metrics.avg_persist_latency_ms.into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        Ok(obj.into())
+    }
+
+    /// Process a temporal sequence and return the resulting hypervector bytes.
+    #[wasm_bindgen(js_name = processSequence)]
+    pub async fn process_sequence(&self, sequence: Array) -> Result<Box<[u8]>, JsValue> {
+        self.framework
+            .validate_sequence_length(sequence.length() as usize)
+            .map_err(to_js_error)?;
+
+        let mut parsed_sequence = Vec::with_capacity(sequence.length() as usize);
+        for item in sequence.iter() {
+            let step = item
+                .dyn_into::<Float32Array>()
+                .map_err(|_| JsValue::from_str("processSequence expects Float32Array items"))?;
+            parsed_sequence.push(step.to_vec());
+        }
+
+        let output = self
+            .framework
+            .process_sequence(&parsed_sequence)
+            .await
+            .map_err(to_js_error)?;
+        Ok(output.to_bytes().into_boxed_slice())
+    }
+
     /// Export all concepts and associations to bytes for in-browser storage.
     #[wasm_bindgen(js_name = exportToBytes)]
     pub async fn export_to_bytes(&self) -> Result<Uint8Array, JsValue> {
@@ -334,66 +435,6 @@ pub fn cosine_similarity(a: &[u8], b: &[u8]) -> Result<f32, JsValue> {
     let hvec_b = HVec10240::from_bytes(b).map_err(to_js_error)?;
 
     Ok(hvec_a.cosine_similarity(&hvec_b))
-}
-
-/// Convert a Concept to a JsValue object
-pub(crate) fn concept_to_js_value(concept: &Concept) -> Result<JsValue, JsValue> {
-    let obj = js_sys::Object::new();
-
-    js_sys::Reflect::set(&obj, &"id".into(), &concept.id.clone().into())
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-
-    js_sys::Reflect::set(
-        &obj,
-        &"vector".into(),
-        &Uint8Array::from(concept.vector.to_bytes().as_slice()),
-    )
-    .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-
-    // Convert metadata HashMap to JS object
-    let metadata_obj = js_sys::Object::new();
-    for (key, value) in &concept.metadata {
-        let value_str = serde_json::to_string(value).map_err(to_js_error)?;
-        let js_value = js_sys::JSON::parse(&value_str)
-            .map_err(|_| JsValue::from_str("failed to parse metadata JSON"))?;
-        js_sys::Reflect::set(&metadata_obj, &key.clone().into(), &js_value)
-            .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-    }
-    js_sys::Reflect::set(&obj, &"metadata".into(), &metadata_obj.into())
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-
-    js_sys::Reflect::set(
-        &obj,
-        &"created_at".into(),
-        &(concept.created_at as f64).into(),
-    )
-    .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-
-    js_sys::Reflect::set(
-        &obj,
-        &"modified_at".into(),
-        &(concept.modified_at as f64).into(),
-    )
-    .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-
-    let expires_at = concept
-        .expires_at
-        .map_or(JsValue::NULL, |v| (v as f64).into());
-    js_sys::Reflect::set(&obj, &"expires_at".into(), &expires_at)
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-
-    let canonical_ids = Array::new();
-    for id in &concept.canonical_concept_ids {
-        canonical_ids.push(&JsValue::from_str(id));
-    }
-    js_sys::Reflect::set(&obj, &"canonical_concept_ids".into(), &canonical_ids.into())
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-
-    Ok(obj.into())
-}
-
-pub(crate) fn to_js_error<E: std::fmt::Display>(error: E) -> JsValue {
-    JsValue::from_str(&error.to_string())
 }
 
 #[wasm_bindgen(typescript_custom_section)]
