+++ src/wasm_ext.rs
@@ -5,123 +5,20 @@
 // Redundant clones are intentional for WASM ownership semantics
 
 #[cfg(target_arch = "wasm32")]
-use js_sys::{Array, Float32Array, Function, Uint8Array};
+use js_sys::{Array, Function, Uint8Array};
 #[cfg(target_arch = "wasm32")]
 use tokio::sync::broadcast::error::RecvError;
-#[cfg(target_arch = "wasm32")]
-use wasm_bindgen::JsCast;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen_futures::spawn_local;
 
 #[cfg(target_arch = "wasm32")]
-use crate::wasm::{WasmFramework, to_js_error};
+use crate::wasm::WasmFramework;
 
 #[cfg(target_arch = "wasm32")]
 #[wasm_bindgen]
 impl WasmFramework {
-    /// Get framework metrics snapshot
-    pub async fn metrics_snapshot(&self) -> Result<JsValue, JsValue> {
-        let metrics = self.framework.metrics_snapshot().await;
-        let obj = js_sys::Object::new();
-        js_sys::Reflect::set(
-            &obj,
-            &"concepts_injected_total".into(),
-            &(metrics.concepts_injected_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"associations_created_total".into(),
-            &(metrics.associations_created_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"probes_total".into(),
-            &(metrics.probes_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"avg_probe_latency_ms".into(),
-            &metrics.avg_probe_latency_ms.into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"cache_hits_total".into(),
-            &(metrics.cache_hits_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"cache_misses_total".into(),
-            &(metrics.cache_misses_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"cache_evictions_total".into(),
-            &(metrics.cache_evictions_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"reservoir_steps_total".into(),
-            &(metrics.reservoir_steps_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"avg_reservoir_step_latency_us".into(),
-            &metrics.avg_reservoir_step_latency_us.into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"reservoir_nodes_active".into(),
-            &(metrics.reservoir_nodes_active as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"persist_ops_total".into(),
-            &(metrics.persist_ops_total as f64).into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        js_sys::Reflect::set(
-            &obj,
-            &"avg_persist_latency_ms".into(),
-            &metrics.avg_persist_latency_ms.into(),
-        )
-        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
-        Ok(obj.into())
-    }
-
-    /// Process a temporal sequence and return the resulting hypervector bytes.
-    #[wasm_bindgen(js_name = processSequence)]
-    pub async fn process_sequence(&self, sequence: Array) -> Result<Box<[u8]>, JsValue> {
-        self.framework
-            .validate_sequence_length(sequence.length() as usize)
-            .map_err(to_js_error)?;
-
-        let mut parsed_sequence = Vec::with_capacity(sequence.length() as usize);
-        for item in sequence.iter() {
-            let step = item
-                .dyn_into::<Float32Array>()
-                .map_err(|_| JsValue::from_str("processSequence expects Float32Array items"))?;
-            parsed_sequence.push(step.to_vec());
-        }
-
-        let output = self
-            .framework
-            .process_sequence(&parsed_sequence)
-            .await
-            .map_err(to_js_error)?;
-        Ok(output.to_bytes().into_boxed_slice())
-    }
     /// Get framework stats
     pub async fn stats(&self) -> Result<JsValue, JsValue> {
         let stats = self.framework.stats().await.map_err(to_js_error)?;
@@ -318,7 +215,7 @@ impl WasmFramework {
             .await
             .map_err(to_js_error)?;
         match concept_opt {
-            Some(concept) => crate::wasm::concept_to_js_value(&concept),
+            Some(concept) => concept_to_js_value(&concept),
             None => Ok(JsValue::NULL),
         }
     }
@@ -331,7 +228,7 @@ impl WasmFramework {
             .rollback_to_version(&id, version as u64)
             .await
             .map_err(to_js_error)?;
-        crate::wasm::concept_to_js_value(&concept)
+        concept_to_js_value(&concept)
     }
 
     /// Export a namespace to bytes for in-browser storage.
@@ -344,6 +241,70 @@ impl WasmFramework {
             .map_err(to_js_error)?;
         Ok(Uint8Array::from(data.as_slice()))
     }
+}
+
+/// Convert a Concept to a JsValue object
+#[cfg(target_arch = "wasm32")]
+pub(crate) fn concept_to_js_value(
+    concept: &crate::singularity::Concept,
+) -> Result<JsValue, JsValue> {
+    let obj = js_sys::Object::new();
+
+    js_sys::Reflect::set(&obj, &"id".into(), &concept.id.clone().into())
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+
+    js_sys::Reflect::set(
+        &obj,
+        &"vector".into(),
+        &Uint8Array::from(concept.vector.to_bytes().as_slice()),
+    )
+    .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+
+    // Convert metadata HashMap to JS object
+    let metadata_obj = js_sys::Object::new();
+    for (key, value) in &concept.metadata {
+        let value_str = serde_json::to_string(value).map_err(to_js_error)?;
+        let js_value = js_sys::JSON::parse(&value_str)
+            .map_err(|_| JsValue::from_str("failed to parse metadata JSON"))?;
+        js_sys::Reflect::set(&metadata_obj, &key.clone().into(), &js_value)
+            .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+    }
+    js_sys::Reflect::set(&obj, &"metadata".into(), &metadata_obj.into())
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+
+    js_sys::Reflect::set(
+        &obj,
+        &"created_at".into(),
+        &(concept.created_at as f64).into(),
+    )
+    .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+
+    js_sys::Reflect::set(
+        &obj,
+        &"modified_at".into(),
+        &(concept.modified_at as f64).into(),
+    )
+    .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+
+    let expires_at = concept
+        .expires_at
+        .map_or(JsValue::NULL, |v| (v as f64).into());
+    js_sys::Reflect::set(&obj, &"expires_at".into(), &expires_at)
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+
+    let canonical_ids = Array::new();
+    for id in &concept.canonical_concept_ids {
+        canonical_ids.push(&JsValue::from_str(id));
+    }
+    js_sys::Reflect::set(&obj, &"canonical_concept_ids".into(), &canonical_ids.into())
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+
+    Ok(obj.into())
+}
+
+#[cfg(target_arch = "wasm32")]
+pub(crate) fn to_js_error<E: std::fmt::Display>(error: E) -> JsValue {
+    JsValue::from_str(&error.to_string())
 }
 
 #[cfg(target_arch = "wasm32")]
