+++ wasm/chaotic_semantic_memory.d.ts
@@ -185,6 +185,19 @@ export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembl
 
 export interface InitOutput {
     readonly memory: WebAssembly.Memory;
+    readonly wasmframework_clear_associations: (a: number, b: number, c: number) => any;
+    readonly wasmframework_concept_count: (a: number) => any;
+    readonly wasmframework_exportNamespaceToBytes: (a: number, b: number, c: number) => any;
+    readonly wasmframework_getVersion: (a: number, b: number, c: number, d: number) => any;
+    readonly wasmframework_inject_text: (a: number, b: number, c: number, d: number, e: number) => any;
+    readonly wasmframework_listVersions: (a: number, b: number, c: number) => any;
+    readonly wasmframework_neighbors: (a: number, b: number, c: number, d: number) => any;
+    readonly wasmframework_on_event: (a: number, b: any) => void;
+    readonly wasmframework_probe_filtered: (a: number, b: number, c: number, d: number, e: number, f: number) => any;
+    readonly wasmframework_probe_text: (a: number, b: number, c: number, d: number) => any;
+    readonly wasmframework_rollbackToVersion: (a: number, b: number, c: number, d: number) => any;
+    readonly wasmframework_stats: (a: number) => any;
+    readonly wasmframework_update_concept_metadata: (a: number, b: number, c: number, d: number, e: number) => any;
     readonly __wbg_wasmframework_free: (a: number, b: number) => void;
     readonly cosine_similarity: (a: number, b: number, c: number, d: number) => [number, number, number];
     readonly encode_text: (a: number, b: number) => [number, number];
@@ -201,9 +214,11 @@ export interface InitOutput {
     readonly wasmframework_importFromBytes: (a: number, b: any, c: number) => any;
     readonly wasmframework_inject_concept: (a: number, b: number, c: number, d: number, e: number) => any;
     readonly wasmframework_inject_concepts: (a: number, b: any, c: any) => any;
+    readonly wasmframework_metrics_snapshot: (a: number) => any;
     readonly wasmframework_new: () => any;
     readonly wasmframework_probe: (a: number, b: number, c: number, d: number) => any;
     readonly wasmframework_probe_batch: (a: number, b: any, c: number) => any;
+    readonly wasmframework_processSequence: (a: number, b: any) => any;
     readonly wasmframework_setNamespace: (a: number, b: number, c: number) => any;
     readonly wasmframework_update_concept: (a: number, b: number, c: number, d: number, e: number) => any;
     readonly wasmframework_bfs: (a: number, b: number, c: number) => any;
@@ -211,24 +226,9 @@ export interface InitOutput {
     readonly wasmframework_probe_with_graph: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number, i: number) => any;
     readonly wasmframework_shortest_path: (a: number, b: number, c: number, d: number, e: number) => any;
     readonly wasmframework_traverse: (a: number, b: number, c: number, d: number, e: number) => any;
-    readonly wasmframework_clear_associations: (a: number, b: number, c: number) => any;
-    readonly wasmframework_concept_count: (a: number) => any;
-    readonly wasmframework_exportNamespaceToBytes: (a: number, b: number, c: number) => any;
-    readonly wasmframework_getVersion: (a: number, b: number, c: number, d: number) => any;
-    readonly wasmframework_inject_text: (a: number, b: number, c: number, d: number, e: number) => any;
-    readonly wasmframework_listVersions: (a: number, b: number, c: number) => any;
-    readonly wasmframework_metrics_snapshot: (a: number) => any;
-    readonly wasmframework_neighbors: (a: number, b: number, c: number, d: number) => any;
-    readonly wasmframework_on_event: (a: number, b: any) => void;
-    readonly wasmframework_probe_filtered: (a: number, b: number, c: number, d: number, e: number, f: number) => any;
-    readonly wasmframework_probe_text: (a: number, b: number, c: number, d: number) => any;
-    readonly wasmframework_processSequence: (a: number, b: any) => any;
-    readonly wasmframework_rollbackToVersion: (a: number, b: number, c: number, d: number) => any;
-    readonly wasmframework_stats: (a: number) => any;
-    readonly wasmframework_update_concept_metadata: (a: number, b: number, c: number, d: number, e: number) => any;
-    readonly wasm_bindgen__convert__closures_____invoke__hebf7812b93ae332a: (a: number, b: number, c: any) => [number, number];
-    readonly wasm_bindgen__convert__closures_____invoke__h0493d3f6782368f7: (a: number, b: number, c: any, d: any) => void;
-    readonly wasm_bindgen__convert__closures_____invoke__h68634416af742c52: (a: number, b: number) => number;
+    readonly wasm_bindgen__convert__closures_____invoke__hbf57a2c43ac1d931: (a: number, b: number, c: any) => [number, number];
+    readonly wasm_bindgen__convert__closures_____invoke__he6e30b5ebcee2b40: (a: number, b: number, c: any, d: any) => void;
+    readonly wasm_bindgen__convert__closures_____invoke__hd6b0f2ad8a8ee62f: (a: number, b: number) => number;
     readonly __wbindgen_malloc: (a: number, b: number) => number;
     readonly __wbindgen_realloc: (a: number, b: number, c: number, d: number) => number;
     readonly __wbindgen_exn_store: (a: number) => void;
